Refactor DropdownMenu component and related files
