--- packages/react/src/components/DropdownMenu/DropdownMenuContext.tsx
@@ -1,11 +1,0 @@
-import { createContext } from 'react';
-
-import type { DropdownMenuProps } from './DropdownMenu';
-
-type DropdownMenuContextType = {
-  size: NonNullable<DropdownMenuProps['size']>;
-};
-
-export const DropdownMenuContext = createContext<DropdownMenuContextType>({
-  size: 'medium',
-});
