--- packages/react/src/components/DropdownMenu/DropdownMenuGroup/index.ts
@@ -1,1 +1,0 @@
-export * from './DropdownMenuGroup';
