--- packages/react/src/components/DropdownMenu/DropdownMenuItem/index.ts
@@ -1,1 +1,0 @@
-export * from './DropdownMenuItem';
