+++ packages/react/src/components/DropdownMenu/index.ts
@@ -1,6 +1,6 @@
 import { DropdownMenu as DropdownRoot } from './DropdownMenu';
-import { DropdownMenuGroup } from './DropdownMenuGroup';
-import { DropdownMenuItem } from './DropdownMenuItem';
+import { DropdownMenuGroup } from './DropdownMenuGroup/DropdownMenuGroup';
+import { DropdownMenuItem } from './DropdownMenuItem/DropdownMenuItem';
 
 type DropdownMenuComponent = typeof DropdownRoot & {
   Group: typeof DropdownMenuGroup;
@@ -15,4 +15,7 @@ DropdownMenu.Item = DropdownMenuItem;
 DropdownMenu.Group.displayName = 'DropdownMenu.Group';
 DropdownMenu.Item.displayName = 'DropdownMenu.Item';
 
+export type { DropdownMenuProps } from './DropdownMenu';
+export type { DropdownMenuGroupProps } from './DropdownMenuGroup/DropdownMenuGroup';
+export type { DropdownMenuItemProps } from './DropdownMenuItem/DropdownMenuItem';
 export { DropdownMenu, DropdownMenuGroup, DropdownMenuItem };
